Replace hand-copied teamVisibility mirror with a real adapter import
teamVisibility.ts was a hand-copied mirror of js/team-visibility.js's
isTeamActive() rule, with a comment asking editors to remember to keep
both files in sync — the same pattern that let js/team-access.js and
the React app's canManageTeamAdmins() drift apart (fixed in #3388).
It hadn't drifted yet, but nothing enforced that.

Route it through a real typed adapter (adapters/legacyTeamVisibility.ts,
matching the ~50 other apps/app/src/lib/adapters/legacy*.ts files) so
there's only one implementation of the rule to keep in sync, and add a
regression guard asserting the adapter still delegates to the legacy
module instead of re-copying the logic.
