+++ apps/app/src/lib/homeLogic.ts
@@ -5,7 +5,7 @@ import {
   type ParentScheduleEvent
 } from './scheduleLogic';
 import type { ParentScheduleChild } from './scheduleService';
-import { isTeamActive } from './teamVisibility';
+import { isTeamActive } from './adapters/legacyTeamVisibility';
 
 export type HomeActionKind = 'rsvp' | 'packet' | 'assignment' | 'rideshare' | 'fee' | 'message';
 export type HomeActionTone = 'amber' | 'blue' | 'emerald' | 'rose' | 'gray';
