+++ tests/unit/app-team-visibility.test.js
@@ -1,8 +1,20 @@
 import { describe, expect, it } from 'vitest';
+import { readFileSync } from 'node:fs';
 
-import { isTeamActive } from '../../apps/app/src/lib/teamVisibility.ts';
+import { isTeamActive } from '../../apps/app/src/lib/adapters/legacyTeamVisibility.ts';
 
 describe('mobile isTeamActive', () => {
+    it('delegates to js/team-visibility.js through the typed adapter instead of hand-copying the rule', () => {
+        // Regression guard: this used to be a hand-copied mirror of the legacy
+        // rule with a comment asking editors to remember to update both files —
+        // exactly the pattern that let js/team-access.js and the React app's
+        // canManageTeamAdmins() drift apart (see PR #3388). Importing through
+        // the adapter means there's only one implementation to keep in sync.
+        const adapterSource = readFileSync('apps/app/src/lib/adapters/legacyTeamVisibility.ts', 'utf8');
+        expect(adapterSource).toContain("from '@legacy/team-visibility.js'");
+    });
+
+
     it('treats missing flags as active (matches legacy default)', () => {
         expect(isTeamActive({})).toBe(true);
         expect(isTeamActive({ active: true })).toBe(true);
